REWRITE
*3.0 Release* - Introducing the new and improved snaibot! Completely
restructured, snaibot now features a modular design, with modules being
toggleable via config (and live!). New in this version is also the
"Choose" module, which can be given a string to be separated by ;'s and
will randomly return one.

While new modules cannot be added live, this should make it
significantly easier to add new features in the future.