Update "source" link
Update "source" link in default config gen. Also generated PyDocs for
both modules and added them to a "docs" folder. Will try to maintain
these in the future, but these should not be treated as the definitive
guides for these. Nothing will replace poking in the raw code for
someone looking to build on top of these things, I'm afraid...