Add Administrative Capabilities and DC Resilience
Cleaned up reconnect functions in main library to better facilitate
handling of server disconnects, including re-registering with NickServ
on reconnect now. Also added a new module for administration. new
commands for Identifying with NickServ manually, joining/leaving
channels, and (de)VOICE-, (de)HOP-, and (de)OPing users in channels.
Also added authentication via config to allow for some key users to be
added to the config to allow use of the above commands in cases where
user might not be in direct control of channel.