*Handler Rewrite*
This update saw revisions to the base pythonircbot library to enable
handlers beyond those in place for messages. Handlers added for join and
parts as well, but more could be added (Unable to add time-based actions
at this time).

Additionally, snaibot was given an upgrade to begin using SQL
capabilities. The first of these allows the bot to now hold requested
user mode settings and, using the new join handler, auto-apply those
modes to users joining.

SQL will likely also be used for the backend on the language and spam
filters in a future update, but are not yet included in this release.