Update to Wiki Module
Rewrote Wiki module to run a proper search using ATLWiki's API and JSON.
Commands are *wiki <term> to return the exact result or top result and
*fullwiki <term> for the bot to query all articles with the search term
in their title. Either command without a term attached will give the
wiki's homepage.

Also tweaked the language and spam filters to only run if bot is hop+ in
the channel. This will allow the bot to sit in non-OP'd channels without
risk of causing spam.